refactor: state 대신 running 추상클래스 상속
